requirements.txt
flask
google-generativeai
pyttsx3